Moved ImageUploader.js under components
